Fixed moving to another level with an awful hack.

+++ src/Player.cpp
@@ -24,7 +24,7 @@ int Player::getCurrentHp(){
 
 void Player::_changeImageBasedOnFacing(){
   std::string newPath;
-  float facingDegrees = getVectorAngleDegrees(_facing);
+  float facingDegrees = static_cast<float>(getVectorAngleDegrees(_facing));
   //Change sprite based on the eight directions.
   if(-22.5 < facingDegrees && facingDegrees < 22.5){
     //Right
